+++ libraries/common/constants/Portals.js
@@ -16,6 +16,7 @@ const GLOBALS = 'globals';
 const ROUTES = 'routes';
 const HEADER = 'header';
 const CONTENT = 'content';
+const CONTENT_LOGIN = 'content-login'
 const HOME = 'home';
 const TERMS = 'terms';
 const PRIVACY = 'privacy';
@@ -34,6 +35,10 @@ export const APP_ROUTES = `${APP}.${ROUTES}`;
 export const PAGE_CONTENT_BEFORE = `${PAGE}.${CONTENT}.${BEFORE}`;
 export const PAGE_CONTENT = `${PAGE}.${CONTENT}`;
 export const PAGE_CONTENT_AFTER = `${PAGE}.${CONTENT}.${AFTER}`;
+
+export const PAGE_CONTENT_LOGIN_BEFORE = `${PAGE}.${CONTENT_LOGIN}.${BEFORE}`;
+export const PAGE_CONTENT_LOGIN = `${PAGE}.${CONTENT_LOGIN}`;
+export const PAGE_CONTENT_AFTER_LOGIN = `${PAGE}.${CONTENT_LOGIN}.${AFTER}`;
 
 export const NAV_MENU_HEADER_BEFORE = `${NAV_MENU}.${HEADER}.${BEFORE}`;
 export const NAV_MENU_HEADER = `${NAV_MENU}.${HEADER}`;