+++ libraries/common/CHANGELOG.md
@@ -1,6 +1,9 @@
 
 # Changelog
 
+## [Unreleased]
+- New `Portal` constant `CONTENT LOGIN` added.
+
 ### [2.4.0](https://github.com/shopgate/pwa-common/compare/v2.3.0...v2.4.0) - February 07, 2018
 #### Added
 - Toast messages supports an action button functionality